Convert the package into a library and command line operation
